Remove `idea` directory from Git
The `.idea/` was tracked on Git. Despite being then added to the `.gitignore` file, it was still present. This commit removes this directory from Git.

The `.idea/` directory is specific to Jetbrains IDEs and contains local files that may change for each user. It's better to not commit it to avoid noise in the Git history.