Maze : Big cleaning, and some refacto
Did a big spring cleanup in MazeMaker.php, made the zone as a variable you set before the generation instead of a parameter you swing around in all function, and [REDACTED] are set before we start the generation. The generation get it directly from the zone. Adapted the code here and there.